feat(java_template): add support for enable-integration-tests and enable-samples maven profiles
In `com.google.cloud:google-cloud-shared-config:0.3.1` integration tests
are skipped by default to reduce friction from contributors that do not
have a service account or access to GCP services. This change updates
our templates to provide instructions on how to enable integration tests
as well as passing the necessary arguments to our builds.

With the incorporation of Code Samples into the project repos we need
new configuration and control in maven to allow building on jdk7 and
excluding samples from release deploy builds. This change adds new build
configs for samples and a new target to build.sh to build samples.