fix: cache default generators
The generator instances can and should be reused within a synthtool
invocation. Currently, to reuse the generator, the synth.py must
instantiate a generator and provide the instance to the gapic_library or
bazel_library helper.

This will allow us to simplify the synth.py files and de-dup the
synth.metadata entries.