fix: don't fail when clean fails
In Ruby's autosynth, some files cannot be cleaned up because, well, I don't know.  But Ruby isn't leaving gigabytes of garbage around so it's not a problem.  So if if git clean is not able to clean up every last file, that should not cause autosynth to file a bug.