fix: fix log path for single-commit, single-branch executions (#613)
Fixes #612 

For the non-multiple commit, non-multiple pull request autosynth runs (Java and Elixir apiaries), we were incorrectly using the base directory as the log file. Now we write to the base_log_directory and optionally expand the path for additional args (used to specify which client we're generating in the Java/Elixir case) and finally specify sponge_log.log as the output file name.