feat(python_samples): Custom path for generation (#695)
Added functionality to be used for generating documentation for Python samples. A demonstration of the documentation this code can render (Using the README.md template under [/python_samples/](https://github.com/googleapis/synthtool/tree/master/synthtool/gcp/templates/python_samples)) is available [here](https://github.com/runargs/python-docs-samples/tree/synth-DEMO/bigtable).

This function will default to generating the README in the current directory if the synth.py file is put in a sample folder, and or the "/samples" folder if it is placed in the top level of a client library repo. A custom path for where to generate may also be specified through the "sample_project_dir" key in .repo.metadata.json