fix: pregenerated code needs to return a copy
Autosynth may not reset the git clone of a given source - other
generators return a temp directory which contains the generated
code.

This fixes the case where autosynth is running the post processing
extra times on the same generated files and the post processing
step is not idempotent (e.g. adding a license header).