feat(bazel): Add clean_build option to synthtool (#995)
In `synth.py`, users will be able to enable build cleaning as follows. This is useful because there is otherwise no way to clear the Bazel cache via the shell.

This should temporarily affect only the PHP microgenerator. Its Bazel rules are implemented in such a way that clobbering the cache is needed if generator changes are made locally, but this will eventually be fixed in the microgenerator. Remote runs will not be affected.
```py
import synthtool.gcp as gcp

gapic = gcp.GAPICBazel()
gapic._clean_build = True
```