chore(owlbot-java): only add modules with version v1 or v2 to parent pom (#1105)
Java-gsuite-addons has an extra set of protos that other protos in the library depend on. These are added to the repo as a separate module `proto-google-apps-script-type-protos`. Only proto-google-cloud-gsuite-addons-v1 depends on this module. But owlbot is currently configured to add all modules to its parent pom. Since the parent do not depend on this extra protos module, it fails the dependency check.
![image](https://user-images.githubusercontent.com/13271327/120238953-42800300-c22b-11eb-9b30-2ffe213d415c.png)


To solve this issue, the post-processor is being configured to add only the modules with a version (v1 or v2) in it.