feat: add env var DEFAULT_BRANCH_PATH
First step to fixing https://github.com/googleapis/repo-automation-bots/issues/2098.

I want to observe the default branch in the first cloud build step,
then pass it to a later step.  However, it can't be passed directly
via an environment variable because steps can't modify their environment
variables and it's not possible to dynamically compose an environment
variable in a cloud-build.yaml.

Therefore, I'll observe the default branch and record it in a file, and
then pass it the file path to the later step.  The contents of the file
may change but the path to the file is constant and can be specified
in cloud-build.yaml.