chore: [java] remove flatten plugin dependency check (#1663)
chore: remove check for flatten plugin

We have had the check for the flatten-maven-plugin in each
Cloud Java client repository. However, the behavior of the plugin
has been stable and its not each repository's responsibility to
assert the plugin's behavior. A new check is going to be added at
the googleapis/java-shared-config repository to assert the
plugin's behavior when we upgrade its version.