docs: fix docs session (#1950)
When the `summary_overview.md` file was added, various `nox -s docs` sessions were complaining as it's not included in any of the toc trees. Since warnings are treated as errors, it fails the builds.

Adding this as a hidden toctree allows the warning to be suppressed, as well as have no effect to the current documentation. Verified that the docs session passes with https://togithub.com/googleapis/python-storage/pull/1252, and then verified locally that docfx sessions are not affected with this change.